Finished adding support for IPv6, may need some optimisation/tiding, but it seems to work.
David did all the best work on it and I realised that PCAP doesn't support tcp[12] style expressions in v6 so was just dropping them.  Have implimented a proper v4 filter with a "or (ip6 and tcp)" and then manually implimented the rest of the filter in C.

The rest of the code is payload based, so having sorted the IP and TCP headers out, the rest just works as is.