Added new --binary (-b) functionality.
The use of this will not be immediately obvious, but this is to output the
fingerprint database as a single binary blob which can be injested by
FingerPrinTLS (feature coming).

This will make performance optimisations beyond just itterating through a
fingerprint array.

The struct output will remain although it will not (hopefully) be used in
FingerPrinTLS in the future, as others may use this for their own code

Approximate output file format:

	# Byte 0		: binary format version
	# Per fingerprint.....
	# Bytes 0-3		: Fingerprint ID
	# Bytes 4-7 	: Desc Length
	# Bytes <above>	: Desc
	# Bytes2 <next>	: record_tls_version;
	# Bytes2 <next>	: tls_version;
	# etc etc etc
	# uint16_t ciphersuite_length
	# uint8_t ciphersuite....
	# uint8_t compression_length
	# uint8_t compression....
	# uint16_t extensions_length
	# uint8_t extensions....
	# uint16_t e_curves_length
	# uint8_t e_curves.....
	# uint16_t sig_alg_length
	# uint8_t sig_alg.....
	# uint16_t ec_point_fmt_length
	# uint8_t ec_point_fmt....