Two main features added:
1 - Fingerprint "database" is not longer a compiled in struct (although the
feature remains in fingerprintout.py for those that still wish to use it).
It has moved to an external file "tlsfp.db".  This is a binary file created by
fingerprintls.py with the -b or --binary switch.  This is becuase the format is
easy to parse in C, JSON... less so.  What does this mean?  Well for the minute
nothing much, but it's laying foudnation for the next steps...  Having an
external file which can be loaded as opposed to compiled in means
adding/removing signatures without recompiling, leaves scope for reload on the
fly, allows me to start writing more efficient search and index options for
performance, dynamically adding fingerprints, etc, etc.

2 - Added an option to drop privileges, so you don't need to keep FingerPrinTLS
running as root.  By using the "-u" switch and specifying a uid (apologies it's
the numeric value for now, username will come soon), FingerPrinTLS will drop
privileges to that of that user as soon as it has parsed arguments, opened file
descriptors, and made the pcap_open_live call.  No packets are parsed until
after privileges are dropped.