Done a sort-of searchable FPDB.
2 dimensional array "search" is used to derive a shorter list to run based on
(masked) ciphersuite length and least significant bit of the TLS version.

These fields were chosen as they are at fixed ofsets and vary meaning that they
are low cost to calculate and populate for the search.  Items in the list
should probably be sorted or further indexed, but this provides (hopefully)
faster search.  Some benchmarking to follow.