Changed how the threading architecture works, no more dispatcher copying packets to threadpool queues....
Now it's threads who use a (blocking) lock request on the pcap handle who then
take a single packet for themselves, take a copy before releasing the lock and
then process.

This removes the need to have a dispatching thread and also removes the need to
manage which queue is more busy, etc.  As threads will only take a packet once
they are free.

This code is not yet optimised, locks are not granualar enough, some may not be
required at all, etc.  And reading a file the code stays in a loop because
pcap_loop looks like same when the end of the file is reached.

However....

The output is now identical on large (race condition causing) pcaps....  So the
results are at least accurate, if not optimised.