Renamed a file and the branch becasue... well... the foray into pthreads has ended.
In this instance threading is not worthwhile, trying to wrangle libpcap into
being useful in a threaded application uses too many cycles to be worthwhile,
not to mention added complexity, etc, etc.  Anyway.... in the process of doing
that I ended up cleaning up a few things so I've retained those changes and am
working on this as a cleanup branch.

Once testing is complete that at least it's functionally equivolent to the main
branch I'll merge then make smaller branches for subsequent cleanup.