Update to add the "-P" flag.
This allows the user to specify a pcap file to which client hello's which do
not match a fingerprint in the database will be stored.  These samples can be
used to verify fingerprints which are generated.  This will also allow me to
being creating a library of samples to use for unit testing later.