Added checks for various offsets pointing beyond the end of the packet.
When incremented past the end of the packet there is a chance (dependant on OS,
compiler, etc) that the pointers could point to addresses in memory which the
fingerprintls process does not have access too or is unallocated.  On my
development system (I believe mostly because of clang) the process just carries
on, I *think* though I need to check that with GCC it may cause a crash.  I do
not believe that anything more serious than a crash can be cause (e.g. code
execution) as the unallocated memory is only read from not written.

If someone who knows more about this than me wants to take a look, I'll happily
take advice on this.